+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## vNEXT
+- Add voucher credit and SRLC manipulation. (#10)
+    - SRLC and iExec poco is mocked.
+    - set voucher credit as VoucherHub is ERC20.
 - Upgrade configuration: (#9)
     - Upgrade dependencies: hardhat, husky, iExec Poco.
     - Ignore mocks in coverage.